sky_20230216
補上ProductDeatail Excel上傳與下載

+++ dbCompanyTest/Controllers/ProductController.cs
@@ -233,6 +233,8 @@ namespace dbCompanyTest.Controllers
             {
                 string excelname = vc.excel.Name;
                 string oldPath = _eviroment.WebRootPath + "/Datas/" + excelname;
+                if (string.IsNullOrEmpty(vc.File_type))
+                    return Json("錯誤_遺失存入資料表代號");
 
 
                 System.IO.File.Delete(oldPath);   //刪掉原本的檔案
@@ -246,14 +248,14 @@ namespace dbCompanyTest.Controllers
                 }
                 using (FileStream fs = new FileStream(path01, FileMode.Open))
                 {
-                    method_uploadEx(fs, vc.excel);
+                    method_uploadEx(fs, vc.excel,vc.File_type);
                 }
             }
             return RedirectToAction("Index");
         }
 
         //將檔案與資料流轉成DataTable並存入資料庫
-        public void method_uploadEx(Stream stream, IFormFile formFile)
+        public void method_uploadEx(Stream stream, IFormFile formFile,string FileType)
         {
             DataTable dataTable = new DataTable();
             IWorkbook wb;
@@ -384,14 +386,91 @@ namespace dbCompanyTest.Controllers
                 stream.Dispose();
                 stream.Close();
             }
+            //將資料存入資料庫
+            DataTable gg = dataTable;
+            switch (FileType)
+            {
+                case "Pro":
+                    method_save_In_Product(dataTable);
+                    break;
+                case "ProDe":
+                    method_save_In_ProDetail(dataTable);
+                    break;
+            }
+        }
 
+        public string method_save_In_ProDetail(DataTable data)
+        {
             //dataTable跑回圈，insert資料至DB
-            foreach (DataRow dr in dataTable.Rows)
+            foreach (DataRow dr in data.Rows)
+            {
+                int _商品編號id = 0;
+                int _商品尺寸id = 0;
+                int _商品顏色id = 0;
+                //dr[8] 與 dr[9] 查詢相應table 回傳可存入的數值
+                if (!string.IsNullOrEmpty(dr[1].ToString()))
+                {
+                    var temp = db.Products.FirstOrDefault(pd => pd.商品名稱 == dr[1].ToString());
+                    if (temp != null)
+                    {
+                        _商品編號id = (int)temp.商品分類id;
+                    }
+                }
+
+                if (!string.IsNullOrEmpty(dr[2].ToString()))
+                {
+                    var temp = db.ProductsSizeDetails.FirstOrDefault(s => s.尺寸種類 == dr[2].ToString());
+                    if (temp != null)
+                    {
+                        _商品尺寸id= temp.商品尺寸id;
+                    }
+                }
+
+                if (!string.IsNullOrEmpty(dr[3].ToString()))
+                {
+                    var temp = db.ProductsColorDetails.FirstOrDefault(s => s.商品顏色種類 == dr[3].ToString());
+                    if (temp != null)
+                    {
+                        _商品顏色id = temp.商品顏色id;
+                    }
+                }
+
+
+                ProductDetail x = new ProductDetail();
+                x.商品編號id = Int32.TryParse(dr[1].ToString(), out int _Id) ? _Id : 0;
+                x.商品尺寸id = Int32.TryParse(dr[2].ToString(), out int _sizeid)? _sizeid:0;
+                x.商品顏色id = Int32.TryParse(dr[3].ToString(), out int _colorid) ? _colorid : 0;
+                x.商品數量 = Int32.TryParse(dr[4].ToString(),out int _count)? _colorid:0;
+                x.商品編號 = dr[1].ToString();
+                x.圖片位置id = Int32.TryParse(dr[6].ToString(), out int _Location) ? _Location : 0;            
+                x.商品是否有貨 = bool.TryParse(dr[7].ToString(), out bool _instock) ? _instock : false;
+                x.商品是否上架 = bool.TryParse(dr[8].ToString(), out bool _onshelves) ? _onshelves : false;
+
+                try
+                {
+                    db.ProductDetails.Add(x);
+                    db.SaveChanges();
+                    //Response.BodyWriter("<script language=javascript>alert('檔案匯入成功');</" + "script>");
+                    return "檔案匯入成功";
+                }
+                catch (Exception ex)
+                {
+                    Console.Write(ex.Message);
+                    return ex.Message;
+                }
+            }
+            return "";
+        }
+
+        public string method_save_In_Product(DataTable data)
+        {
+            //dataTable跑回圈，insert資料至DB
+            foreach (DataRow dr in data.Rows)
             {
                 int _分類id = 0;
                 int _鞋種id = 0;
-                //dr[8] 與 dr[9] 查詢相應table 回傳可存入的數值
-                if (!string.IsNullOrEmpty(dr[8].ToString()))
+                //dr[9] 與 dr[10] 查詢相應table 回傳可存入的數值
+                if (!string.IsNullOrEmpty(dr[9].ToString()))
                 {
                     var temp = db.ProductsTypeDetails.FirstOrDefault(pd => pd.商品分類名稱 == dr[9].ToString());
                     if (temp != null)
@@ -400,7 +479,7 @@ namespace dbCompanyTest.Controllers
                     }
                 }
 
-                if (!string.IsNullOrEmpty(dr[9].ToString()))
+                if (!string.IsNullOrEmpty(dr[10].ToString()))
                 {
                     var temp = db.商品鞋種s.FirstOrDefault(s => s.鞋種 == dr[10].ToString());
                     if (temp != null)
@@ -410,7 +489,7 @@ namespace dbCompanyTest.Controllers
                 }
 
 
-                Product x = new Product();
+                Product x = new Product();                
                 x.上架時間 = dr[1].ToString();
                 x.商品名稱 = dr[2].ToString();
                 x.商品價格 = Convert.ToDecimal(double.TryParse(dr[3].ToString(), out double _price) ? _price : 0);
@@ -425,16 +504,18 @@ namespace dbCompanyTest.Controllers
 
                 try
                 {
-
                     db.Products.Add(x);
                     db.SaveChanges();
                     //Response.BodyWriter("<script language=javascript>alert('檔案匯入成功');</" + "script>");
+                    return "檔案匯入成功";
                 }
                 catch (Exception ex)
                 {
                     Console.Write(ex.Message);
+                    return ex.Message;
                 }
             }
+            return "";
         }
 
         //下載
